+++ function/python/brightics/function/classification/decision_tree_classification.py
@@ -8,8 +8,8 @@ from brightics.common.repr import dict2MD
 from brightics.function.utils import _model_dict
 from brightics.common.groupby import _function_by_group
 from brightics.common.utils import check_required_parameters
-from brightics.function.validation import validate
-from brightics.function.validation import greater_than_or_equal_to
+from brightics.common.validation import validate
+from brightics.common.validation import greater_than_or_equal_to
 
 
 def decision_tree_classification_train(table, group_by=None, **params):
