+++ function/python/brightics/function/classification/logistic_regression.py
@@ -7,8 +7,8 @@ from brightics.common.repr import pandasDF2MD
 from brightics.function.utils import _model_dict
 from brightics.common.groupby import _function_by_group
 from brightics.common.utils import check_required_parameters
-from brightics.function.validation import raise_runtime_error
-from brightics.function.validation import raise_error
+from brightics.common.validation import raise_runtime_error
+from brightics.common.validation import raise_error
 import sklearn.utils as sklearn_utils
