+++ function/python/brightics/function/classification/svm_classification.py
@@ -5,7 +5,7 @@ from brightics.function.utils import _model_dict
 from brightics.common.repr import BrtcReprBuilder, strip_margin, pandasDF2MD, dict2MD
 from brightics.common.groupby import _function_by_group
 from brightics.common.utils import check_required_parameters
-from brightics.function.validation import validate, greater_than, \
+from brightics.common.validation import validate, greater_than, \
     raise_runtime_error
 import sklearn.utils as sklearn_utils
