+++ function/python/brightics/function/classification/xgb_classification.py
@@ -7,7 +7,7 @@ from brightics.function.utils import _model_dict
 from brightics.common.groupby import _function_by_group
 from brightics.common.utils import check_required_parameters
 import numpy as np
-from brightics.function.validation import validate, greater_than_or_equal_to
+from brightics.common.validation import validate, greater_than_or_equal_to
 
 
 def xgb_classification_train(table, group_by=None, **params):
