+++ function/python/brightics/function/clustering/kmeans.py
@@ -9,7 +9,7 @@ from sklearn.metrics.cluster.unsupervised import silhouette_score, \
 from brightics.function.utils import _model_dict
 from brightics.common.groupby import _function_by_group
 from brightics.common.utils import check_required_parameters
-from brightics.function.validation import validate, greater_than_or_equal_to, \
+from brightics.common.validation import validate, greater_than_or_equal_to, \
     greater_than, all_elements_greater_than, raise_runtime_error
