+++ function/python/brightics/function/manipulation/__init__.py
@@ -5,7 +5,7 @@ import numpy as np
 import math
 from brightics.common.groupby import _function_by_group
 from brightics.common.utils import check_required_parameters
-from brightics.function.validation import validate, greater_than_or_equal_to, raise_error, require_param
+from brightics.common.validation import validate, greater_than_or_equal_to, raise_error, require_param
 
 
 def filter(table, query):
