+++ function/python/brightics/function/regression/decision_tree_regression.py
@@ -7,7 +7,7 @@ from brightics.function.utils import _model_dict
 from sklearn.tree.export import export_graphviz
 from brightics.common.groupby import _function_by_group
 from brightics.common.utils import check_required_parameters
-from brightics.function.validation import validate, greater_than_or_equal_to
+from brightics.common.validation import validate, greater_than_or_equal_to
 
 
 def decision_tree_regression_train(table, group_by=None, **params):
