+++ function/python/brightics/function/regression/glm.py
@@ -3,7 +3,7 @@ from brightics.common.repr import BrtcReprBuilder, strip_margin
 from brightics.function.utils import _model_dict
 from brightics.common.groupby import _function_by_group
 from brightics.common.utils import check_required_parameters
-from brightics.function.validation import raise_runtime_error
+from brightics.common.validation import raise_runtime_error
 
 
 def glm_train(table, group_by=None, **params):
