+++ function/python/brightics/function/statistics/cross_table.py
@@ -2,7 +2,7 @@ from brightics.common.repr import BrtcReprBuilder, strip_margin, pandasDF2MD
 from brightics.function.utils import _model_dict
 from brightics.common.groupby import _function_by_group
 from brightics.common.utils import check_required_parameters
-from brightics.function.validation import raise_runtime_error
+from brightics.common.validation import raise_runtime_error
 
 import numpy as np
 import pandas as pd
