+++ function/python/brightics/function/statistics/pairplot.py
@@ -3,7 +3,7 @@ import seaborn as sns
 from brightics.common.repr import BrtcReprBuilder
 from brightics.common.groupby import _function_by_group
 from brightics.common.utils import check_required_parameters
-from brightics.function.validation import validate, greater_than
+from brightics.common.validation import validate, greater_than
 
 
 def pairplot(table, group_by=None, **params):
