+++ function/python/brightics/function/transform/sample.py
@@ -1,6 +1,6 @@
 from brightics.common.groupby import _function_by_group
 from brightics.common.utils import check_required_parameters
-from brightics.function.validation import validate, greater_than_or_equal_to
+from brightics.common.validation import validate, greater_than_or_equal_to
 
 
 def random_sampling(table, group_by=None, **params):
